+++ fgmetric/converters/__init__.py
@@ -1,8 +1,10 @@
+from fgmetric.converters._bool_tokens import BoolTokens
 from fgmetric.converters._counter_pivot_table import CounterPivotTable
 from fgmetric.converters._delimited_list import DelimitedList
 from fgmetric.converters._null_sentinels import NullSentinels
 
 __all__ = [
+    "BoolTokens",
     "CounterPivotTable",
     "DelimitedList",
     "NullSentinels",